Hero split in two views. Nos mobile view added.
