+++ src/components/molecules/AlbumHero/AlbumHero.sass
@@ -1,23 +1,9 @@
-@import '@/style/functions'
-@import '@/style/colors'
-@import '@/style/textHelper'
+@import '@/style/breakpoints'
 
 .AlbumHero
-	width: 100vw
-	background-repeat: no-repeat
-	background-size: cover
-	background-position: center
-	height: 100vw
-	display: flex
-	flex-direction: column
-	justify-content: flex-end
-	align-items: center
-	padding: pixelsToRem(24)
-	box-sizing: border-box
-	-moz-box-sizing: border-box
-	-webkit-box-sizing: border-box
-	&__title
-		font-size: min(8vw, 6rem)
-		color: $color-white
-		word-break: break-word
-		@include paragraphEllipsis(2)
+	&__mobile
+		@media screen and (min-width: $break-computer)
+			display: none
+	&__desktop
+		@media screen and (max-width: $break-computer)
+			display: none
