+++ src/components/molecules/AlbumHero/AlbumHero.vue
@@ -1,36 +1,22 @@
 <template>
-  <div
-    class="AlbumHero"
-    :style="backgroundImage"
-  >
-    <h1 class="AlbumHero__title">
-      {{ title }}
-    </h1>
+  <div>
+    <div class="AlbumHero__mobile">
+      <AlbumHeroMobile
+        img="https://picsum.photos/200"
+        title="Fallling"
+      />
+    </div>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import { linearGradientImage } from '../../../helpers/imageHelper';
+import AlbumHeroMobile from './AlbumHeroMobile/AlbumHeroMobile.vue';
 
 export default defineComponent({
   name: 'AlbumHero',
-  props: {
-    img: {
-      type: String,
-      required: true,
-    },
-    title: {
-      type: String,
-      required: true,
-    },
-  },
-  setup(props) {
-    const backgroundImage = linearGradientImage(props.img);
-
-    return {
-      backgroundImage,
-    };
+  components: {
+    AlbumHeroMobile,
   },
 });
 </script>
